requirements.txt
pytesseract==0.3.10
pymupdf
nltk==3.7
numpy
pandas
regex
more_itertools
scikit-learn
transformers
glob2
plotly_express
jellyfish
nbformat
sentencepiece
torch
sacremoses
tensorflow
pyyaml
easyocr
pillow